Fix existing tests to account for history wrapped stored state.
